+++ src/App.css
@@ -1,6 +1,7 @@
 body {
   background-color: #282c34;
   color: aliceblue;
+  font-weight: 500;
 }
 
 .app {
