+++ src/components/User/UserProfile.css
@@ -33,3 +33,31 @@ font-size: medium;
 line-height: 1.5;
 width: 50vw;
 }
+
+.user-add-photo {
+  --c:  white;
+  
+  font-weight: 600;
+  box-shadow: 0 0 0 .1em inset var(--c); 
+  --_g: linear-gradient(var(--c) 0 0) no-repeat;
+  background: 
+    var(--_g) calc(var(--_p,0%) - 100%) 0%,
+    var(--_g) calc(200% - var(--_p,0%)) 0%,
+    var(--_g) calc(var(--_p,0%) - 100%) 100%,
+    var(--_g) calc(200% - var(--_p,0%)) 100%;
+  background-size: 50.5% calc(var(--_p,0%)/2 + .5%);
+  outline-offset: .1em;
+  transition: background-size .4s, background-position 0s .4s;
+}
+.user-add-photo:hover {
+  --alt: #282c34;
+  --_p: 100%;
+
+  color: var(--alt);
+  transition: background-position .4s, background-size 0s;
+}
+.user-add-photo:active {
+  box-shadow: 0 0 9e9q inset #0009; 
+  background-color: var(--c);
+  color: #fff;
+}
